put read & read_list in same file
