+++ src/contents/clothing/mod.rs
@@ -2,5 +2,4 @@ pub mod clothing_error;
 pub mod delete;
 pub mod page_view;
 pub mod read;
-pub mod read_list;
 pub mod write;
