+++ src/contents/clothing/read.rs
@@ -1,4 +1,5 @@
-use crate::{firestore::read::read_firestore, *};
+use crate::{firestore::read::{read_firestore, read_list_firestore}, *};
+use firestore_db_and_auth::documents;
 
 pub async fn clothing_read(
     session: Session,
@@ -10,7 +11,7 @@ pub async fn clothing_read(
 
     let document_id = String::from(&params.document_id);
 
-    let read_result: DemoDTOClothing = match read_firestore(session, &auth, &document_id).await {
+    let read_result: DTOClothing = match read_firestore(session, &auth, &document_id).await {
         Err(e) => return Err(e.into()),
         Ok(dto) => dto,
     };
@@ -29,14 +30,41 @@ pub async fn clothing_read(
     Ok(HttpResponse::Ok().content_type("text/html").body(view))
 }
 
-#[derive(Serialize, Deserialize)]
-pub struct FormParamsDbRead {
-    document_id: String,
+pub async fn clothing_read_list(
+    session: Session,
+    tmpl: web::Data<Tera>,
+) -> actix_web::Result<HttpResponse, super::clothing_error::ReadListError> {
+    //ServiceSession reference is included in documents::List, so get auth outside of
+    //read_list_firestore function.
+    let cred = Credentials::from_file("firebase-service-account.json").unwrap();
+    let auth = ServiceSession::new(cred).unwrap();
+
+    let read_list_result: documents::List<DTOClothing, _> =
+        match read_list_firestore(session, &auth).await {
+            Err(e) => return Err(e.into()),
+            Ok(dto_list) => dto_list,
+        };
+
+    for doc_result in read_list_result {
+        let (doc, _metadata) = doc_result?;
+        println!("{doc:?}");
+    }
+
+    let context = Context::new();
+    let view = tmpl
+        .render("clothing.html", &context)
+        .map_err(error::ErrorInternalServerError)?;
+    Ok(HttpResponse::Ok().content_type("text/html").body(view))
 }
 
-#[derive(Serialize, Deserialize)]
-struct DemoDTOClothing {
+#[derive(Serialize, Deserialize, Debug)]
+struct DTOClothing {
     brand: String,
     year: u32,
     month: u32,
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct FormParamsDbRead {
+    document_id: String,
 }
